python/requirements.txt
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
Pillow==10.1.0
